Fix mine danger logic: TOGGLING is safe, not dangerous
Critical bug fix identified by user:
- TOGGLING (state=2): Player is overlapping, mine is safe
- TOGGLED (state=0): Mine is deadly, instant kill

Changes:
- Fixed test expectations in test_mine_state_processor.py
- Corrected documentation in subtask_rewards.py
- Extracted MineAwareSubtaskContext to separate file (mine_aware_context.py)
  to keep subtask_policies.py under 500 line limit
- Updated TASK_2.3_IMPLEMENTATION_SUMMARY.md with correct mechanics
