+++ .env
@@ -1,0 +1,12 @@
+# NPP-RL Environment Configuration
+# Add current directory to Python path (will be expanded when sourced)
+PYTHONPATH=${PYTHONPATH}:${PWD}
+
+# Training defaults (CPU-only)
+NPP_RL_NUM_ENVS=4
+NPP_RL_TOTAL_TIMESTEPS=10000
+
+# Logging
+NPP_RL_LOG_LEVEL=INFO
+NPP_RL_LOG_DIR=./training_logs
+# Current project path: /workspace/npp-rl
